+++ src/FIFOCalculator/FIFOCalculator.csproj
@@ -32,7 +32,7 @@
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
     <PackageReference Include="Serilog.Sinks.XUnit" Version="3.0.5" />
     <PackageReference Include="System.Text.Json" Version="7.0.2" />
-    <PackageReference Include="Zafiro.Avalonia" Version="3.0.9" />
+    <PackageReference Include="Zafiro.Avalonia" Version="3.0.12" />
     <PackageReference Include="Zafiro.Core" Version="3.2.33" />
   </ItemGroup>
