+++ src/FIFOCalculator/Views/DataEntryView.axaml
@@ -1,0 +1,28 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:views="clr-namespace:FIFOCalculator.Views"
+             xmlns:viewModels="clr-namespace:FIFOCalculator.ViewModels"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:CompileBindings="True"
+             x:DataType="viewModels:DataEntryViewModel"
+             x:Class="FIFOCalculator.Views.DataEntryView">
+    <DockPanel>
+        <UniformGrid DockPanel.Dock="Bottom" Rows="1" DataContext="{Binding LoadStoreViewModel}">
+            <UniformGrid.Styles>
+                <Style Selector="Button">
+                    <Setter Property="HorizontalAlignment" Value="Stretch"></Setter>
+                </Style>
+            </UniformGrid.Styles>
+            <Button Command="{Binding New}" Content="New" />
+            <Button Command="{Binding Open}" Content="Open" />
+            <Button Command="{Binding Save}" Content="Save" />
+        </UniformGrid>
+        <UniformGrid Columns="1">
+            <views:EntryEditorView DataContext="{Binding Inputs}" />
+            <views:EntryEditorView DataContext="{Binding Outputs}" />
+        </UniformGrid>
+    </DockPanel>
+
+</UserControl>
